requirements.txt
Flask==3.1.0
google-generativeai==0.8.4
joblib==1.4.2
numpy==2.2.3
gunicorn==23.0.0
Werkzeug==3.1.3
scikit-learn==1.6.1
flask_cors==5.0.1